BACKEND/requirements.txt
Django==4.1.4
django-cors-headers==3.13.0
django-rest-framework
PyYAML==6.0.0
uritemplate==4.1.1
requests==2.28.2
django-admin-list-filter-dropdown==1.0.3
django-cachalot==2.5.2
django-configurations==2.4.0
django-cursor-pagination==0.2.1
psycopg2==2.9.5
django-filter==22.1.0
python-dotenv==0.21.1
